quantum/ibm/requirements.txt
# requirements.txt
--constraint https://qisk.it/1-0-constraints

qiskit
qiskit-ibm-runtime
python-dotenv
matplotlib
